= Moved `gmStorage` traits to their own file
