+++ src/ECS/mod.rs
@@ -5,7 +5,10 @@ use std::collections::{BTreeMap, HashMap};
 use super::*;
 
 mod comp;
+mod storage;
+
 use comp::*;
+use storage::*;
 
 pub trait gmRes: Any{
     fn new() -> Self;
@@ -187,20 +190,6 @@ impl gmDispatchStage{
         for SYS in self.inner.iter_mut(){
             SYS.executeNow(IN_world);
         }
-    }
-}
-
-pub trait gmStorage<T>: Any{
-    fn new() -> Self where Self: Sized;
-    fn insert(&mut self, IN_id: gmID, IN_item: T);
-    fn remove(&mut self, IN_id: gmID) -> Option<T>;
-}
-pub trait gmStorageDrop: Any{
-    fn drop(&mut self, IN_id: gmID);
-}
-impl<T: 'static> gmStorageDrop for dyn gmStorage<T>{
-    fn drop(&mut self, IN_id: gmID) {
-        self.remove(IN_id);
     }
 }
